feat(forum_conversation): move can_certify_post method to forum_conversation app
