+++ lacommunaute/forum_conversation/shortcuts.py
@@ -1,5 +1,6 @@
 from django.db.models import Count, Exists, OuterRef, Prefetch, Q, QuerySet
 
+from lacommunaute.forum.enums import Kind as Forum_Kind
 from lacommunaute.forum_conversation.models import Post, Topic
 from lacommunaute.forum_upvote.models import UpVote
 from lacommunaute.users.models import User
@@ -28,3 +29,11 @@ def get_posts_of_a_topic_except_first_one(topic: Topic, user: User) -> QuerySet[
             upvotes_count=Count("upvotes"),
         )
     return qs.order_by("created")
+
+
+def can_certify_post(forum, user):
+    return (
+        user.is_authenticated
+        and forum.kind == Forum_Kind.PUBLIC_FORUM
+        and (user.groups.filter(forum=forum).exists() or user.is_staff)
+    )
