+++ lacommunaute/forum_conversation/views.py
@@ -12,8 +12,7 @@ from lacommunaute.forum.models import Forum
 from lacommunaute.forum_conversation.enums import Filters
 from lacommunaute.forum_conversation.forms import PostForm, TopicForm
 from lacommunaute.forum_conversation.models import Topic
-from lacommunaute.forum_conversation.shortcuts import get_posts_of_a_topic_except_first_one
-from lacommunaute.forum_upvote.shortcuts import can_certify_post
+from lacommunaute.forum_conversation.shortcuts import can_certify_post, get_posts_of_a_topic_except_first_one
 
 
 logger = logging.getLogger(__name__)
